polishing

+++ JugMeasure.Core/Capacity.cs
@@ -2,12 +2,12 @@
 {
     public struct Capacity
     {
-        public Capacity(ushort v)
+        public Capacity(uint v)
         {
             this.Value = v;
         }
 
-        public ushort Value { get; }
+        public uint Value { get; }
 
         public static bool operator <(Capacity x, Capacity y)
         {
@@ -23,12 +23,12 @@
         {
             if (x < y)
                 return new Capacity(0);
-            return new Capacity((ushort) (x.Value - y.Value));
+            return new Capacity((uint) (x.Value - y.Value));
         }
 
         public static Capacity operator +(Capacity x, Capacity y)
         {
-            return new Capacity((ushort)(x.Value + y.Value));
+            return new Capacity((uint)(x.Value + y.Value));
         }
 
         public static bool operator ==(Capacity x, Capacity y)
